+++ infra/forge-cli/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @granite-js/forge-cli
 
+## 0.1.34-next.25
+
+### Patch Changes
+
+- fix
+- Updated dependencies
+  - @granite-js/deployment-manager@0.1.34-next.25
+  - @granite-js/plugin-core@0.1.34-next.25
+
 ## 0.1.34-next.24
 
 ### Patch Changes