+++ packages/naver-map/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @granite-js/naver-map
 
+## 0.1.34-next.25
+
+### Patch Changes
+
+- fix
+
 ## 0.1.34-next.24
 
 ### Patch Changes