+++ packages/plugin-hermes/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @granite-js/plugin-hermes
 
+## 0.1.34-next.25
+
+### Patch Changes
+
+- fix
+- Updated dependencies
+  - @granite-js/plugin-core@0.1.34-next.25
+  - @granite-js/utils@0.1.34-next.25
+
 ## 0.1.34-next.24
 
 ### Patch Changes